refactor(tile-entities): rename directory
